requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
python-dotenv>=1.0.1
httpx>=0.27.0
# torch is required, but precise rocm installation is environment dependent 
# e.g.: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/rocm6.0
torch>=2.0.0
